appease stricter gcc warnings

+++ src/StateSerializer.h
@@ -15,14 +15,14 @@ public:
 
 		auto* subElement = root->createNewChildElement("numerator_coefficients");
 		juce::StringArray bCoeffStrArray;
-		for (int i = 0; i < zeroCoeffs.size(); i++)
-			bCoeffStrArray.add(juce::String(zeroCoeffs[static_cast<size_t>(i)]));
+		for (size_t i = 0; i < zeroCoeffs.size(); i++)
+			bCoeffStrArray.add(juce::String(zeroCoeffs[i]));
 		subElement->addTextElement(bCoeffStrArray.joinIntoString(", "));
 
 		subElement = root->createNewChildElement("denominator_coefficients");
 		juce::StringArray aCoeffStrArray;
-		for (int i = 0; i < poleCoeffs.size(); i++)
-			aCoeffStrArray.add(juce::String(poleCoeffs[static_cast<size_t>(i)]));
+		for (size_t i = 0; i < poleCoeffs.size(); i++)
+			aCoeffStrArray.add(juce::String(poleCoeffs[i]));
 		subElement->addTextElement(aCoeffStrArray.joinIntoString(", "));
 
 		return root;
